requirements.txt
flake8==3.9.2
flake8-docstrings==1.6.0
pytest==6.2.5
python-telegram-bot==13.7
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
isort==5.12.0
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
schedule==1.2.0
six==1.16.0
tornado==6.2
tzdata==2023.3
tzlocal==4.3
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
openai==0.27.6
requests==2.30.0
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
